Clean up methods in joruney plan view.
Extract these out to relevant structs / views
